requirements.txt
fiona==1.9.5
geopandas==0.14.3
pandas==2.2.0
shapely==2.0.3
pyproj==3.6.1
rasterio==1.3.9
rasterstats==0.19.0
scikit-image==0.22.0
scipy==1.12.0
segmentation-models-pytorch==0.3.3
tensorboard==2.16.2
transformers==4.37.2
ultralytics==8.2.2
scikit-learn==1.4.1.post1
torchmetrics==0.11.4
kornia==0.7.1